Converts source data to float64 for precision
Increases the precision of the source data conversion by using `np.float64` instead of `np.float32`.

This improves the accuracy of subsequent calculations that rely on this converted data.

+++ src/processing/converters/source.py
@@ -22,7 +22,7 @@ class SourceConverter:
 
     def convert(self, data: np.ndarray) -> np.ndarray:
         """Convert float32 data to 32-bit signed integer range."""
-        float_data = data.astype(np.float32, copy=False)
+        float_data = data.astype(np.float64, copy=False)
         # Scale to 32-bit signed integer range: [-2^31, 2^31-1]
         scaled = np.clip(np.rint(float_data * 2147483648.0), -2147483648, 2147483647)
         return scaled.astype(np.int32)
